backend/django/requirements.txt
# Django core
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
dj-database-url==2.1.0
python-decouple==3.8
whitenoise==6.9.0

# ASGI server
daphne==4.2.0

# Database
psycopg2-binary>=2.9,<3.0

# Background tasks
dramatiq[redis]==1.18.0
redis==6.2.0

# Auth
PyJWT==2.9.0

# OTP delivery via WhatsApp Cloud API (sync HTTP from Django views)
requests>=2.31,<3.0

# Firebase (FCM push notifications)
firebase-admin==6.8.0

# Google Maps
googlemaps==4.10.0

# Google Cloud Storage (media uploads)
google-cloud-storage==3.1.0

# Image processing
pillow==11.2.1

# Prometheus monitoring
django-prometheus==2.3.1

# API docs
drf-yasg==1.21.10

# Payments (Paynow Zimbabwe)
paynow==1.0.8

# AWS SES (email sending)
boto3>=1.35,<2.0

# Timezone
pytz==2025.2
